refactor(combinatorics): isolate recurrence contracts
